Moves ZukiniStepContext back into core zukini project (I believe there was a name collision when it was in the Zukini Project before.) This also gets the errors to go away for me.
