+++ Zukini/Zukini.csproj
@@ -55,6 +55,7 @@
     <Compile Include="PropertyAlreadyExistsException.cs" />
     <Compile Include="PropertyBucket.cs" />
     <Compile Include="PropertyNotFoundException.cs" />
+    <Compile Include="ZukiniStepContext.cs" />
   </ItemGroup>
   <ItemGroup>
     <None Include="App.config" />
